+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 Records breaking changes from major version bumps
 
+## 14.0.0
+
+PR: [#116](https://github.com/alphagov/digitalmarketplace-apiclient/pull/116)
+
+`search_services` method on the search-api-client has been generalized to `search` and now takes a `doc_type` argument.
+This allows more than just services to be searched.
+
+`aggregate` method on the search-api-client has been generalized to `aggregate` to allow for aggregating
+briefs too. It takes an extra argument, `doc_type`.
+
+Old:
+```
+search_api_client.aggregate_services(index=index, q=q, aggregations=aggregations)
+search_api_client.search_services(index, q=q, page=page, id_only=id_only, **filters)
+```
+
+New:
+```
+search_api_client.aggregate(index=index, doc_type='services', q=q, aggregations=aggregations)
+search_api_client.search(index, doc_type, q=q, page=page, id_only=id_only, **filters)
+```
+
 ## 13.0.0
 
 PR: [#115](https://github.com/alphagov/digitalmarketplace-apiclient/pull/115)